requirements.txt
asgiref
Django
sqlparse
gunicorn
psycopg2-binary>=2.8.6
python-dateutil
